+++ posthog/hogql_queries/experiments/test/experiment_query_runner/test_experiment_query_runner.py
@@ -1,26 +1,17 @@
 import json
 from typing import cast
 from django.test import override_settings
-from pytest import mark
 from posthog.constants import ExperimentNoResultsErrorKeys
 from posthog.hogql_queries.experiments.experiment_query_runner import ExperimentQueryRunner
 from posthog.hogql_queries.experiments.test.experiment_query_runner.utils import (
     create_standard_group_test_events,
-    create_data_warehouse_table,
 )
+from posthog.hogql_queries.experiments.test.experiment_query_runner.base import ExperimentQueryRunnerBaseTest
 from posthog.models.action.action import Action
 from posthog.models.cohort.cohort import Cohort
-from posthog.models.feature_flag.feature_flag import FeatureFlag
 from posthog.models.group_type_mapping import GroupTypeMapping
 from posthog.models.group.util import create_group
 from rest_framework.exceptions import ValidationError
-from posthog.settings import (
-    OBJECT_STORAGE_ACCESS_KEY_ID,
-    OBJECT_STORAGE_BUCKET,
-    OBJECT_STORAGE_ENDPOINT,
-    OBJECT_STORAGE_SECRET_ACCESS_KEY,
-    XDIST_SUFFIX,
-)
 from posthog.schema import (
     ActionsNode,
     EventsNode,
@@ -32,245 +23,24 @@ from posthog.schema import (
     ExperimentSignificanceCode,
     ExperimentVariantFunnelsBaseStats,
     ExperimentVariantTrendsBaseStats,
-    PersonsOnEventsMode,
     ExperimentFunnelMetric,
     ExperimentMeanMetric,
     PropertyOperator,
 )
 from posthog.test.base import (
-    APIBaseTest,
-    ClickhouseTestMixin,
     _create_event,
     _create_person,
-    create_person_id_override_by_distinct_id,
     flush_persons_and_events,
     snapshot_clickhouse_queries,
 )
 from freezegun import freeze_time
-from django.utils import timezone
 from datetime import datetime, timedelta
 from posthog.test.test_journeys import journeys_for
-from posthog.models.experiment import Experiment
 from parameterized import parameterized
-from boto3 import resource
-from botocore.config import Config
-from posthog.warehouse.models.join import DataWarehouseJoin
-
-TEST_BUCKET = "test_storage_bucket-posthog.hogql.experiments.queryrunner" + XDIST_SUFFIX
 
 
 @override_settings(IN_UNIT_TESTING=True)
-class TestExperimentQueryRunner(ClickhouseTestMixin, APIBaseTest):
-    def teardown_method(self, method) -> None:
-        s3 = resource(
-            "s3",
-            endpoint_url=OBJECT_STORAGE_ENDPOINT,
-            aws_access_key_id=OBJECT_STORAGE_ACCESS_KEY_ID,
-            aws_secret_access_key=OBJECT_STORAGE_SECRET_ACCESS_KEY,
-            config=Config(signature_version="s3v4"),
-            region_name="us-east-1",
-        )
-        bucket = s3.Bucket(OBJECT_STORAGE_BUCKET)
-        bucket.objects.filter(Prefix=TEST_BUCKET).delete()
-
-    def create_feature_flag(self, key="test-experiment"):
-        return FeatureFlag.objects.create(
-            name=f"Test experiment flag: {key}",
-            key=key,
-            team=self.team,
-            filters={
-                "groups": [{"properties": [], "rollout_percentage": None}],
-                "multivariate": {
-                    "variants": [
-                        {
-                            "key": "control",
-                            "name": "Control",
-                            "rollout_percentage": 50,
-                        },
-                        {
-                            "key": "test",
-                            "name": "Test",
-                            "rollout_percentage": 50,
-                        },
-                    ]
-                },
-            },
-            created_by=self.user,
-        )
-
-    def create_experiment(
-        self,
-        name="test-experiment",
-        feature_flag=None,
-        start_date=None,
-        end_date=None,
-    ):
-        if feature_flag is None:
-            feature_flag = self.create_feature_flag(name)
-        if start_date is None:
-            start_date = timezone.now()
-        else:
-            start_date = timezone.make_aware(start_date)  # Make naive datetime timezone-aware
-        if end_date is None:
-            end_date = timezone.now() + timedelta(days=14)
-        elif end_date is not None:
-            end_date = timezone.make_aware(end_date)  # Make naive datetime timezone-aware
-        return Experiment.objects.create(
-            name=name,
-            team=self.team,
-            feature_flag=feature_flag,
-            start_date=start_date,
-            end_date=end_date,
-            exposure_criteria=None,
-        )
-
-    def create_data_warehouse_table_with_usage(self):
-        table_data = [
-            {"id": "1", "ds": "2023-01-01", "userid": "user_control_0", "usage": "1000"},
-            {"id": "2", "ds": "2023-01-02", "userid": "user_test_1", "usage": "500"},
-            {"id": "3", "ds": "2023-01-03", "userid": "user_test_2", "usage": "750"},
-            {"id": "4", "ds": "2023-01-04", "userid": "internal_test_1", "usage": "100000"},
-            {"id": "5", "ds": "2023-01-06", "userid": "user_test_3", "usage": "800"},
-            {"id": "6", "ds": "2023-01-07", "userid": "user_extra", "usage": "900"},
-        ]
-
-        columns = {
-            "id": "String",
-            "ds": "Date",
-            "userid": "String",
-            "usage": "String",
-        }
-
-        table_name = "usage"
-
-        create_data_warehouse_table(self.team, table_name, table_data, columns)
-
-        return table_name
-
-    def create_data_warehouse_table_with_subscriptions(self):
-        subscription_table_name = "subscriptions"
-        subscription_columns = {
-            "subscription_id": "String",
-            "subscription_created_at": "DateTime64(3, 'UTC')",
-            "subscription_customer_id": "String",
-            "subscription_amount": "Int64",
-        }
-        subscription_table_data = [
-            {
-                "subscription_id": "1",
-                "subscription_created_at": datetime(2023, 1, 2),
-                "subscription_customer_id": "1",
-                "subscription_amount": 100,
-            },
-            {
-                "subscription_id": "2",
-                "subscription_created_at": datetime(2023, 1, 3),
-                "subscription_customer_id": "2",
-                "subscription_amount": 50,
-            },
-            {
-                "subscription_id": "3",
-                "subscription_created_at": datetime(2023, 1, 4),
-                "subscription_customer_id": "3",
-                "subscription_amount": 75,
-            },
-            {
-                "subscription_id": "4",
-                "subscription_created_at": datetime(2023, 1, 5),
-                "subscription_customer_id": "4",
-                "subscription_amount": 80,
-            },
-            {
-                "subscription_id": "5",
-                "subscription_created_at": datetime(2023, 1, 6),
-                "subscription_customer_id": "5",
-                "subscription_amount": 90,
-            },
-        ]
-        create_data_warehouse_table(self.team, subscription_table_name, subscription_table_data, subscription_columns)
-
-        customer_table_name = "customers"
-        customer_columns = {
-            "customer_id": "String",
-            "customer_created_at": "DateTime64(3, 'UTC')",
-            "customer_name": "String",
-            "customer_email": "String",
-        }
-        customer_table_data = [
-            {
-                "customer_id": "1",
-                "customer_created_at": datetime(2023, 1, 1),
-                "customer_name": "John Doe",
-                "customer_email": "john.doe@example.com",
-            },
-            {
-                "customer_id": "2",
-                "customer_created_at": datetime(2023, 1, 2),
-                "customer_name": "Jane Doe",
-                "customer_email": "jane.doe@example.com",
-            },
-            {
-                "customer_id": "3",
-                "customer_created_at": datetime(2023, 1, 3),
-                "customer_name": "John Smith",
-                "customer_email": "john.smith@example.com",
-            },
-            {
-                "customer_id": "4",
-                "customer_created_at": datetime(2023, 1, 6),
-                "customer_name": "Jane Smith",
-                "customer_email": "jane.smith@example.com",
-            },
-            {
-                "customer_id": "5",
-                "customer_created_at": datetime(2023, 1, 7),
-                "customer_name": "John Doe Jr",
-                "customer_email": "john.doejr@example.com",
-            },
-        ]
-        create_data_warehouse_table(self.team, customer_table_name, customer_table_data, customer_columns)
-
-        DataWarehouseJoin.objects.create(
-            team=self.team,
-            source_table_name=subscription_table_name,
-            source_table_key="subscription_customer_id",
-            joining_table_name=customer_table_name,
-            joining_table_key="customer_id",
-            field_name="subscription_customer",
-        )
-
-        return subscription_table_name
-
-    def create_standard_test_events(self, feature_flag):
-        """
-        Creates a standard set of events that can be reused across multiple tests
-        """
-
-        feature_flag_property = f"$feature/{feature_flag.key}"
-
-        for variant, purchase_count in [("control", 6), ("test", 8)]:
-            for i in range(10):
-                _create_person(distinct_ids=[f"user_{variant}_{i}"], team_id=self.team.pk)
-                _create_event(
-                    team=self.team,
-                    event="$feature_flag_called",
-                    distinct_id=f"user_{variant}_{i}",
-                    timestamp="2020-01-02T12:00:00Z",
-                    properties={
-                        feature_flag_property: variant,
-                        "$feature_flag_response": variant,
-                        "$feature_flag": feature_flag.key,
-                    },
-                )
-                if i < purchase_count:
-                    _create_event(
-                        team=self.team,
-                        event="purchase",
-                        distinct_id=f"user_{variant}_{i}",
-                        timestamp="2020-01-02T12:01:00Z",
-                        properties={feature_flag_property: variant, "amount": 10 if i < 2 else ""},
-                    )
-
+class TestExperimentQueryRunner(ExperimentQueryRunnerBaseTest):
     @freeze_time("2020-01-01T12:00:00Z")
     @snapshot_clickhouse_queries
     def test_query_runner_funnel_metric(self):
@@ -718,50 +488,6 @@ class TestExperimentQueryRunner(ClickhouseTestMixin, APIBaseTest):
         self.assertEqual(control_variant.count, 60)
         self.assertEqual(test_variant.count, 120)
 
-    @freeze_time("2020-01-01T12:00:00Z")
-    @snapshot_clickhouse_queries
-    def test_query_runner_group_aggregation_funnel_metric(self):
-        feature_flag = self.create_feature_flag()
-        feature_flag.filters["aggregation_group_type_index"] = 0
-        feature_flag.save()
-        experiment = self.create_experiment(feature_flag=feature_flag)
-
-        metric = ExperimentFunnelMetric(
-            series=[
-                EventsNode(event="purchase"),
-            ],
-        )
-
-        experiment_query = ExperimentQuery(
-            experiment_id=experiment.id,
-            kind="ExperimentQuery",
-            metric=metric,
-        )
-
-        experiment.metrics = [metric.model_dump(mode="json")]
-        experiment.save()
-
-        create_standard_group_test_events(self.team, feature_flag)
-
-        flush_persons_and_events()
-
-        query_runner = ExperimentQueryRunner(query=experiment_query, team=self.team)
-        result = query_runner.calculate()
-
-        self.assertEqual(len(result.variants), 2)
-
-        control_variant = cast(
-            ExperimentVariantFunnelsBaseStats, next(variant for variant in result.variants if variant.key == "control")
-        )
-        test_variant = cast(
-            ExperimentVariantFunnelsBaseStats, next(variant for variant in result.variants if variant.key == "test")
-        )
-
-        self.assertEqual(control_variant.success_count, 2)
-        self.assertEqual(test_variant.success_count, 3)
-        self.assertEqual(control_variant.failure_count, 0)
-        self.assertEqual(test_variant.failure_count, 0)
-
     @snapshot_clickhouse_queries
     def test_query_runner_group_aggregation_data_warehouse_mean_metric(self):
         feature_flag = self.create_feature_flag()
@@ -864,281 +590,6 @@ class TestExperimentQueryRunner(ClickhouseTestMixin, APIBaseTest):
         self.assertEqual(test_result.absolute_exposure, 2)
         self.assertEqual(control_result.count, 1)
         self.assertEqual(test_result.count, 3)
-
-    @parameterized.expand(
-        [
-            ###
-            # PERSON_ID_OVERRIDE_PROPERTIES_ON_EVENTS
-            ###
-            [
-                "person_id_override_properties_on_events_no_filter",
-                PersonsOnEventsMode.PERSON_ID_OVERRIDE_PROPERTIES_ON_EVENTS,
-                None,
-                {
-                    "control_success": 1,
-                    "control_failure": 0,
-                    "test_success": 1,
-                    "test_failure": 0,
-                },
-            ],
-            [
-                "person_id_override_properties_on_events_filter_earlierevent",
-                PersonsOnEventsMode.PERSON_ID_OVERRIDE_PROPERTIES_ON_EVENTS,
-                {
-                    "key": "email",
-                    "value": "@earlierevent.com",
-                    "operator": "not_icontains",
-                    "type": "person",
-                },
-                {
-                    "control_success": 1,
-                    "control_failure": 0,
-                    "test_success": 1,
-                    "test_failure": 0,
-                },
-            ],
-            [
-                "person_id_override_properties_on_events_filter_laterevent",
-                PersonsOnEventsMode.PERSON_ID_OVERRIDE_PROPERTIES_ON_EVENTS,
-                {
-                    "key": "email",
-                    "value": "@laterevent.com",
-                    "operator": "not_icontains",
-                    "type": "person",
-                },
-                {
-                    "control_success": 1,
-                    "control_failure": 0,
-                    "test_success": 0,
-                    "test_failure": 1,
-                },
-            ],
-            ###
-            # PERSON_ID_OVERRIDE_PROPERTIES_JOINED
-            ###
-            [
-                "person_id_override_properties_joined_no_filter",
-                PersonsOnEventsMode.PERSON_ID_OVERRIDE_PROPERTIES_JOINED,
-                None,
-                {
-                    "control_success": 1,
-                    "control_failure": 0,
-                    "test_success": 1,
-                    "test_failure": 0,
-                },
-            ],
-            [
-                "person_id_override_properties_joined_filter_earlierevent",
-                PersonsOnEventsMode.PERSON_ID_OVERRIDE_PROPERTIES_JOINED,
-                {
-                    "key": "email",
-                    "value": "@earlierevent.com",
-                    "operator": "not_icontains",
-                    "type": "person",
-                },
-                {
-                    "control_success": 1,
-                    "control_failure": 0,
-                    "test_success": 1,
-                    "test_failure": 0,
-                },
-            ],
-            [
-                "person_id_override_properties_joined_filter_laterevent",
-                PersonsOnEventsMode.PERSON_ID_OVERRIDE_PROPERTIES_JOINED,
-                {
-                    "key": "email",
-                    "value": "@laterevent.com",
-                    "operator": "not_icontains",
-                    "type": "person",
-                },
-                None,
-            ],
-            ###
-            # PERSON_ID_NO_OVERRIDE_PROPERTIES_ON_EVENTS
-            ###
-            [
-                "person_id_no_override_properties_on_events_no_filter",
-                PersonsOnEventsMode.PERSON_ID_NO_OVERRIDE_PROPERTIES_ON_EVENTS,
-                None,
-                {
-                    "control_success": 1,
-                    "control_failure": 0,
-                    "test_success": 1,
-                    "test_failure": 1,
-                },
-            ],
-            [
-                "person_id_no_override_properties_on_events_filter_earlierevent",
-                PersonsOnEventsMode.PERSON_ID_NO_OVERRIDE_PROPERTIES_ON_EVENTS,
-                {
-                    "key": "email",
-                    "value": "@earlierevent.com",
-                    "operator": "not_icontains",
-                    "type": "person",
-                },
-                {
-                    "control_success": 1,
-                    "control_failure": 0,
-                    "test_success": 1,
-                    "test_failure": 0,
-                },
-            ],
-            [
-                "person_id_no_override_properties_on_events_filter_laterevent",
-                PersonsOnEventsMode.PERSON_ID_NO_OVERRIDE_PROPERTIES_ON_EVENTS,
-                {
-                    "key": "email",
-                    "value": "@laterevent.com",
-                    "operator": "not_icontains",
-                    "type": "person",
-                },
-                {
-                    "control_success": 1,
-                    "control_failure": 0,
-                    "test_success": 0,
-                    "test_failure": 1,
-                },
-            ],
-        ]
-    )
-    @snapshot_clickhouse_queries
-    @freeze_time("2020-01-01T12:00:00Z")
-    def test_query_runner_with_persons_on_events_mode(self, name, persons_on_events_mode, filters, expected_results):
-        feature_flag = self.create_feature_flag()
-        experiment = self.create_experiment(
-            feature_flag=feature_flag,
-            start_date=datetime(2020, 1, 1),
-            end_date=datetime(2020, 1, 31),
-        )
-        experiment.stats_config = {"version": 2}
-        experiment.save()
-
-        feature_flag_property = f"$feature/{feature_flag.key}"
-
-        experiment_query = ExperimentQuery(
-            experiment_id=experiment.id,
-            kind="ExperimentQuery",
-            metric=ExperimentFunnelMetric(
-                series=[
-                    EventsNode(event="purchase"),
-                ],
-            ),
-        )
-        experiment.exposure_criteria = {"filterTestAccounts": True}
-        experiment.metrics = [{"type": "primary", "query": experiment_query.model_dump()}]
-        experiment.save()
-
-        ## Control isn't affected by the filter
-        _create_person(distinct_ids=["user_control_1"], team_id=self.team.pk)
-        _create_event(
-            team=self.team,
-            event="$feature_flag_called",
-            distinct_id="user_control_1",
-            timestamp="2020-01-02T12:00:00Z",
-            properties={
-                "$feature_flag": feature_flag.key,
-                feature_flag_property: "control",
-                "$feature_flag_response": "control",
-            },
-        )
-        _create_event(
-            team=self.team,
-            event="purchase",
-            distinct_id="user_control_1",
-            timestamp="2020-01-02T12:01:00Z",
-            properties={feature_flag_property: "control"},
-        )
-
-        ## Test is tied to person on events mode
-        _create_person(
-            distinct_ids=["person_id_1_distinct_id_1"],
-            properties={"email": "person_id_1@earlierevent.com"},
-            team_id=self.team.pk,
-        )
-        _create_event(
-            team=self.team,
-            event="$feature_flag_called",
-            distinct_id="person_id_1_distinct_id_1",
-            timestamp="2020-01-02T12:00:00Z",
-            properties={
-                "$feature_flag": feature_flag.key,
-                feature_flag_property: "test",
-                "$feature_flag_response": "test",
-            },
-        )
-        _create_person(
-            distinct_ids=["person_id_1_distinct_id_2"],
-            properties={"email": "person_id_1@laterevent.com"},
-            team_id=self.team.pk,
-        )
-        _create_event(
-            team=self.team,
-            event="$feature_flag_called",
-            distinct_id="person_id_1_distinct_id_2",
-            timestamp="2020-01-02T12:01:00Z",
-            properties={
-                "$feature_flag": feature_flag.key,
-                feature_flag_property: "test",
-                "$feature_flag_response": "test",
-            },
-        )
-        _create_event(
-            team=self.team,
-            event="purchase",
-            distinct_id="person_id_1_distinct_id_2",
-            timestamp="2020-01-02T12:02:00Z",
-            properties={feature_flag_property: "test"},
-        )
-        create_person_id_override_by_distinct_id("person_id_1_distinct_id_1", "person_id_1_distinct_id_2", self.team.pk)
-
-        flush_persons_and_events()
-
-        self.team.modifiers = {"personsOnEventsMode": persons_on_events_mode}
-        if filters:
-            self.team.test_account_filters = [filters]
-        self.team.save()
-
-        query_runner = ExperimentQueryRunner(query=experiment_query, team=self.team)
-        if expected_results is None:
-            with self.assertRaises(ValidationError) as context:
-                query_runner.calculate()
-
-            if "person_id_override_properties_joined_filter_laterevent" in name:
-                expected_errors = json.dumps(
-                    {
-                        ExperimentNoResultsErrorKeys.NO_EXPOSURES: False,
-                        ExperimentNoResultsErrorKeys.NO_CONTROL_VARIANT: False,
-                        ExperimentNoResultsErrorKeys.NO_TEST_VARIANT: True,
-                    }
-                )
-            else:
-                expected_errors = json.dumps(
-                    {
-                        ExperimentNoResultsErrorKeys.NO_EXPOSURES: True,
-                        ExperimentNoResultsErrorKeys.NO_CONTROL_VARIANT: True,
-                        ExperimentNoResultsErrorKeys.NO_TEST_VARIANT: True,
-                    }
-                )
-            self.assertEqual(cast(list, context.exception.detail)[0], expected_errors)
-        else:
-            result = query_runner.calculate()
-
-            self.assertEqual(len(result.variants), 2)
-            control_variant = cast(
-                ExperimentVariantFunnelsBaseStats, next(v for v in result.variants if v.key == "control")
-            )
-            test_variant = cast(ExperimentVariantFunnelsBaseStats, next(v for v in result.variants if v.key == "test"))
-
-            self.assertEqual(
-                {
-                    "control_success": int(control_variant.success_count),
-                    "control_failure": int(control_variant.failure_count),
-                    "test_success": int(test_variant.success_count),
-                    "test_failure": int(test_variant.failure_count),
-                },
-                expected_results,
-            )
 
     @freeze_time("2020-01-01T12:00:00Z")
     @snapshot_clickhouse_queries
@@ -1254,78 +705,6 @@ class TestExperimentQueryRunner(ClickhouseTestMixin, APIBaseTest):
         self.assertEqual(test_variant.count, 5.0)
         # In the new query runner, the exposure value is the same as the absolute exposure value
         self.assertEqual(test_variant.exposure, 2.0)
-
-    @mark.skip("Funnel metrics on data warehouse tables are not supported yet")
-    @snapshot_clickhouse_queries
-    def test_query_runner_data_warehouse_funnel_metric(self):
-        # table_name = self.create_data_warehouse_table_with_usage()
-
-        feature_flag = self.create_feature_flag()
-        experiment = self.create_experiment(
-            feature_flag=feature_flag, start_date=datetime(2023, 1, 1), end_date=datetime(2023, 1, 31)
-        )
-        experiment.stats_config = {"version": 2}
-        experiment.save()
-
-        feature_flag_property = f"$feature/{feature_flag.key}"
-
-        metric = ExperimentFunnelMetric(
-            # TODO: fix this once supported
-            # source=ExperimentDataWarehouseNode(
-            #     table_name=table_name,
-            #     events_join_key="properties.$user_id",
-            #     data_warehouse_join_key="userid",
-            #     timestamp_field="ds",
-            # ),
-            series=[
-                EventsNode(event="purchase"),
-            ],
-        )
-        experiment_query = ExperimentQuery(
-            experiment_id=experiment.id,
-            kind="ExperimentQuery",
-            metric=metric,
-        )
-        experiment.exposure_criteria = {"filterTestAccounts": False}
-        experiment.metrics = [metric.model_dump(mode="json")]
-        experiment.save()
-
-        # Populate exposure events
-        for variant, count in [("control", 7), ("test", 9)]:
-            for i in range(count):
-                _create_event(
-                    team=self.team,
-                    event="$feature_flag_called",
-                    distinct_id=f"distinct_{variant}_{i}",
-                    properties={
-                        "$feature_flag_response": variant,
-                        feature_flag_property: variant,
-                        "$feature_flag": feature_flag.key,
-                        "$user_id": f"user_{variant}_{i}",
-                        "$group_0": "my_awesome_group",
-                    },
-                    timestamp=datetime(2023, 1, i + 1),
-                )
-
-        flush_persons_and_events()
-
-        query_runner = ExperimentQueryRunner(query=experiment_query, team=self.team)
-        with freeze_time("2023-01-07"):
-            result = query_runner.calculate()
-
-        self.assertEqual(len(result.variants), 2)
-
-        control_result = cast(
-            ExperimentVariantFunnelsBaseStats, next(variant for variant in result.variants if variant.key == "control")
-        )
-        test_result = cast(
-            ExperimentVariantFunnelsBaseStats, next(variant for variant in result.variants if variant.key == "test")
-        )
-
-        self.assertEqual(control_result.success_count, 1)
-        self.assertEqual(test_result.success_count, 3)
-        self.assertEqual(control_result.failure_count, 6)
-        self.assertEqual(test_result.failure_count, 6)
 
     @snapshot_clickhouse_queries
     def test_query_runner_data_warehouse_count_metric(self):
@@ -2718,185 +2097,3 @@ class TestExperimentQueryRunner(ClickhouseTestMixin, APIBaseTest):
         # Verify the exposure counts (users who have been exposed to the variant)
         self.assertEqual(control_variant.absolute_exposure, 1)  # Only user_control
         self.assertEqual(test_variant.absolute_exposure, 1)  # Only user_test
-
-    @freeze_time("2024-01-01T12:00:00Z")
-    @snapshot_clickhouse_queries
-    def test_query_runner_funnel_metric_config(self):
-        feature_flag = self.create_feature_flag()
-        experiment = self.create_experiment(feature_flag=feature_flag)
-        experiment.stats_config = {"version": 2}
-        experiment.save()
-
-        ff_property = f"$feature/{feature_flag.key}"
-
-        # Create test data using journeys
-        journeys_for(
-            {
-                # User with first step only
-                "user_control_1": [
-                    {
-                        "event": "$feature_flag_called",
-                        "timestamp": "2024-01-02T12:00:00",
-                        "properties": {
-                            "$feature_flag_response": "control",
-                            ff_property: "control",
-                            "$feature_flag": feature_flag.key,
-                        },
-                    },
-                    {
-                        "event": "$pageview",
-                        "timestamp": "2024-01-02T12:01:00",
-                        "properties": {
-                            ff_property: "control",
-                        },
-                    },
-                ],
-                # User with first and second step completed
-                "user_control_2": [
-                    {
-                        "event": "$feature_flag_called",
-                        "timestamp": "2024-01-02T12:00:00",
-                        "properties": {
-                            "$feature_flag_response": "control",
-                            ff_property: "control",
-                            "$feature_flag": feature_flag.key,
-                        },
-                    },
-                    {
-                        "event": "$pageview",
-                        "timestamp": "2024-01-02T12:01:00",
-                        "properties": {
-                            ff_property: "control",
-                        },
-                    },
-                    {
-                        "event": "purchase",
-                        "timestamp": "2024-01-03T12:02:00",
-                        "properties": {
-                            ff_property: "control",
-                        },
-                    },
-                ],
-                # User with second step only
-                "user_control_3": [
-                    {
-                        "event": "$feature_flag_called",
-                        "timestamp": "2024-01-03T12:00:00",
-                        "properties": {
-                            "$feature_flag_response": "control",
-                            ff_property: "control",
-                            "$feature_flag": feature_flag.key,
-                        },
-                    },
-                    {
-                        "event": "purchase",
-                        "timestamp": "2024-01-03T12:02:00",
-                        "properties": {
-                            ff_property: "control",
-                        },
-                    },
-                ],
-                # User with only first step completed
-                "user_test_1": [
-                    {
-                        "event": "$feature_flag_called",
-                        "timestamp": "2024-01-02T12:00:00",
-                        "properties": {
-                            "$feature_flag_response": "test",
-                            ff_property: "test",
-                            "$feature_flag": feature_flag.key,
-                        },
-                    },
-                    {
-                        "event": "$pageview",
-                        "timestamp": "2024-01-02T12:01:00",
-                        "properties": {
-                            ff_property: "test",
-                        },
-                    },
-                ],
-                # User with whole funnel completed
-                "user_test_2": [
-                    {
-                        "event": "$feature_flag_called",
-                        "timestamp": "2024-01-02T12:00:00",
-                        "properties": {
-                            "$feature_flag_response": "test",
-                            ff_property: "test",
-                            "$feature_flag": feature_flag.key,
-                        },
-                    },
-                    {
-                        "event": "$pageview",
-                        "timestamp": "2024-01-03T12:01:00",
-                        "properties": {
-                            ff_property: "test",
-                        },
-                    },
-                    {
-                        "event": "purchase",
-                        "timestamp": "2024-01-03T12:02:00",
-                        "properties": {
-                            ff_property: "test",
-                        },
-                    },
-                ],
-                # User with only feature flag and purchase, no pageview. Should be excluded.
-                "user_test_3": [
-                    {
-                        "event": "$feature_flag_called",
-                        "timestamp": "2024-01-02T12:00:00",
-                        "properties": {
-                            "$feature_flag_response": "test",
-                            ff_property: "test",
-                            "$feature_flag": feature_flag.key,
-                        },
-                    },
-                    {
-                        "event": "purchase",
-                        "timestamp": "2024-01-03T12:02:00",
-                        "properties": {
-                            ff_property: "test",
-                        },
-                    },
-                ],
-            },
-            self.team,
-        )
-
-        flush_persons_and_events()
-
-        metric = ExperimentFunnelMetric(
-            series=[
-                EventsNode(event="$pageview"),
-                EventsNode(event="purchase"),
-            ],
-        )
-
-        experiment_query = ExperimentQuery(
-            experiment_id=experiment.id,
-            kind="ExperimentQuery",
-            metric=metric,
-        )
-
-        experiment.metrics = [metric.model_dump(mode="json")]
-        experiment.save()
-
-        flush_persons_and_events()
-
-        query_runner = ExperimentQueryRunner(query=experiment_query, team=self.team)
-        result = query_runner.calculate()
-
-        self.assertEqual(len(result.variants), 2)
-
-        control_variant = cast(
-            ExperimentVariantFunnelsBaseStats, next(variant for variant in result.variants if variant.key == "control")
-        )
-        test_variant = cast(
-            ExperimentVariantFunnelsBaseStats, next(variant for variant in result.variants if variant.key == "test")
-        )
-
-        self.assertEqual(control_variant.success_count, 1)
-        self.assertEqual(control_variant.failure_count, 2)
-        self.assertEqual(test_variant.success_count, 1)
-        self.assertEqual(test_variant.failure_count, 2)
